feat:-refactor: I agree a few funtionalitis and refactor
I make a few new funtionalitis for securitiy and one more. I refactor ,a make a big hook Im planing to refactor a litle more and make 3 or 4 hooks and finally i make a few change in the Style. All this chances are in the frontend of the app, I didnt chance to much in the backend